Added screen for when user is logged in already, plus logout

+++ src/app/App.jsx
@@ -18,7 +18,7 @@ const App = () => {
     };
 
     const deleteUserIfDontKeep = (e) => {
-      deleteUser();
+      if (checkIfUserExists() && !getUserAttribute('keepLogin')) deleteUser();
     };
 
     window.addEventListener('beforeunload', warnUser);
